fix(coin-evm): address Copilot review on getBlock dedup PR
- Move `dropRootTraceDuplicates` (and its `signatureOf` helper) out of
  `logic/getBlock.ts` into a new `logic/rootTraceDedup.ts` so the helper is
  no longer re-exported via `logic/index.ts`. It's an implementation detail
  of `getBlock`; tests import it directly from the new module.
- Extract `NON_VALUE_TRANSFER_CALL_TYPES` into `adapters/nonValueTransferCallTypes.ts`
  and import it from both `adapters/etherscan.ts` and `adapters/blockOperations.ts`.
  Not re-exported via `adapters/index.ts`. Removes the duplicated declaration +
  "keep in sync" note.
- Restore the stricter `BigInt(it.value) <= 0n` rejection in
  `isInternalTransactionValid` (accidentally loosened to `=== 0n` in the
  previous commit). Avoids accepting malformed negative `value` payloads.
- Update the `mergeInternalTransactions` docstring to match the actual dedup
  key (`(address, peer, amount)` on native ops only; ERC20/721/1155 not
  touched because they come from receipt logs).
- Reorder imports in `getBlock.test.ts` so the external type-only import
  comes first, and restore an SC-interaction test that was accidentally
  dropped when the branch was rebased onto the newer develop.
